complete ssh command functionality
